+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Settings dialog.
 - Configurable device name.
+- Basic video attachment support.
 
 ### Fixed